started another cleanup with minor fixes

+++ X10D.Performant/src/BooleanExtensions/BooleanExtensions.cs
@@ -31,29 +31,29 @@
         /// </summary>
         /// <param name="value">A <see cref="bool"/>.</param>
         /// <returns>1 if <paramref name="value"/> is <see langword="true"/>, or 0 otherwise.</returns>
-        public static long ToInt64(this bool value) => value ? 1 : 0;        
-        
+        public static long ToInt64(this bool value) => value ? 1 : 0;
+
         /// <summary>
         ///     Gets the value of this <see cref="bool"/> as represented by <see cref="decimal"/>.
         /// </summary>
         /// <param name="value">A <see cref="bool"/>.</param>
         /// <returns>1 if <paramref name="value"/> is <see langword="true"/>, or 0 otherwise.</returns>
         public static decimal ToDecimal(this bool value) => value ? 1 : 0;
-        
+
         /// <summary>
         ///     Gets the value of this <see cref="bool"/> as represented by <see cref="double"/>.
         /// </summary>
         /// <param name="value">A <see cref="bool"/>.</param>
         /// <returns>1 if <paramref name="value"/> is <see langword="true"/>, or 0 otherwise.</returns>
         public static double ToDouble(this bool value) => value ? 1 : 0;
-        
+
         /// <summary>
         ///     Gets the value of this <see cref="bool"/> as represented by <see cref="float"/>.
         /// </summary>
         /// <param name="value">A <see cref="bool"/>.</param>
         /// <returns>1 if <paramref name="value"/> is <see langword="true"/>, or 0 otherwise.</returns>
-        public static double ToFloat(this bool value) => value ? 1 : 0;
-        
+        public static float ToSingle(this bool value) => value ? 1 : 0;
+
         /// <summary>
         ///     Gets the value of this <see cref="bool"/> as represented by <see cref="sbyte"/>.
         /// </summary>
